Move newly-added Webauthn tests out of gocheck
